Use uint for lb IDs

+++ Clubber.Discord/Modules/Stats.cs
@@ -69,7 +69,7 @@ public class Stats : ExtendedModulebase<SocketCommandContext>
 			return;
 		}
 
-		await ShowStats((uint)ddUser.LeaderboardId, user);
+		await ShowStats(ddUser.LeaderboardId, user);
 	}
 
 	private async Task ShowStats(uint lbId, SocketGuildUser? user)
